2017 から 2019 への変更に伴う修正 GAME_OBJECT を GameObject に変更

+++ AegisEngine/Sprite_Animation.cpp
@@ -234,7 +234,7 @@ void SPRITE_ANIMATION::Update(float delta_time)
 		FrameCount = 0.0f;
 	}
 
-	PatternCount = (int)(Age / WaitFrame);
+	PatternCount = WORD(Age / WaitFrame);
 
 	if ((Pattern_Max_X * Pattern_Max_Y) <= PatternCount)
 	{
